requirements.txt
flask
pynacl
python-dotenv
ffmpeg
discord
asyncio
